Adding the daybook-push command.
This command lets you push a service asset into S3 while specifying the service name and version, naming your object properly to work with daybook-pull.  It can take advantage of the same configuration as daybook-pull if the credentials have both read *and* write permissions, so it's a nice little companion tool.